app/requirements.txt
fastapi==0.121.1
mysql_connector_repackaged==0.3.1
python-dotenv==1.2.1
uvicorn==0.38.0
